% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_coor_ws.R
\name{all_coor_ws}
\alias{all_coor_ws}
\title{all_coor_ws() retrieve weather station within a region ( State/County )}
\usage{
all_coor_ws(ghcnd, geoid, type)
}
\arguments{
\item{ghcnd}{String that refers to the Global Historical Climate Network (daily) dataset}

\item{geoid}{FIPS number from census tables (tiger files)}

\item{type}{Variable name under investigation i.e. TMAX, TMIN, PRCP}
}
\value{
The function returns a data frame. When it fails to retrieve information, it returns 1.
}
\description{
The function retrieves all NOAA weather station (ws) data available in a region ( State/County ) related
to the weather variable under investigation (type).
}
\examples{
ghcnd <- 'GHCND'
geoid <- '12087'
type  <- 'PRCP'
all_coor_ws( ghcnd, geoid, type)
}
\keyword{NOAA,}
\keyword{rnoaa}
\keyword{stations,}
\keyword{weather}
