% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ws2pgdb.R
\name{createVoronoi}
\alias{createVoronoi}
\title{createVoronoi creates a Voronoi tessellation}
\usage{
createVoronoi(tableName, the_geom, ogc_fid, scale)
}
\arguments{
\item{tableName}{The name of existing table within the Postgres database.}

\item{the_geom}{The column name with the geometries of the generating points.}

\item{ogc_fid}{The column name with the unique identifiers for each row element (point).}

\item{scale}{The percentage of buffer distance (how far the expand outward)}
}
\value{
The function returns the name of the voronoi table that has been created.
}
\description{
This function creates a table with polygone geometries of a Voronoi tessellation in a Postgres database.
}
\note{
In https://gist.github.com/djq/4714788 you will find the query code to be loaded into the database.
The r_voronoi(text,text,text) function is needed to create the Voronoi tessellation.
createVoronoi() function exploits this code, but a small modification was required.
The buffer_distance has been made flexible in contrast to the original code that sets the buffer distance to 10%.
The r_voronoi(text,text,text) has been replaced by the r_voronoi_scale(text,text,text,float),
and TYPE was renamed as the r_voronoi_scale_type. In the original code, search for the section with the comment
"calculate an approprate buffer distance (~10%):" and replace (0.10) for (arg4). That should suffice
to make createVoronoi() to work as explained in the example.
}
\examples{
ghcnd <- 'GHCND'
geoid <- '40'
type  <- 'TMAX'
sufix <- 'create_voronoi_example'
tableName <- all_coor_ws_2_pgdb( ghcnd, geoid, type, sufix )
createVoronoi(tableName, "geom", "ogc_fid", 0.25)
}

