% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dengue_model_conn.R
\name{dengue_model_conn}
\alias{dengue_model_conn}
\title{dengue_model_conn() habia una vez ...}
\usage{
dengue_model_conn(tableData, disease, conn, degree)
}
\arguments{
\item{tableData}{is the name of the table containing weather data ( temperature )}

\item{disease}{is the name of the vector-borne disease (i.e dengue, malaria, chikungunya, etc)}

\item{conn}{is an open connectino to read/write to a pgdb}

\item{degree}{is the order at one wants to adjusts the polynomio (for dengue degree=5) has the lowest norm}
}
\value{
returns the table name with evaluation of the model over the available data.
}
\description{
dengue_model_conn() habia una vez ...
}
\examples{
file   <- base::paste(Sys.getenv("HOME"), "/","pg_config.yml", sep="")
config <- yaml::yaml.load_file( file )
driver <- "PostgreSQL"
drv    <- RPostgres::Postgres()
conn   <- RPostgres::dbConnect(drv, host= config$dbhost, port= config$dbport, dbname= config$dbname, user= config$dbuser, password= config$dbpwd)
ghcnd <- 'GHCND'
geoid <- '12087'
type  <- 'TMAX'
disease <- 'dengue'
degree  <- '5'
stations <- as.data.frame( all_coor_ws( ghcnd, geoid, type) )
span <- '2'
tableData <- ws_metadata_span_2_pgdb( geoid, type, stations, span)
tableDisease <- base::paste(tableData,"_dengue",sep="")
q    <- base::paste("select r_table_exists('", tableDisease,"')", sep="")
res   <- RPostgres::dbSendQuery(conn, q)
exists <- RPostgres::dbFetch(res)
RPostgres::dbClearResult(res)
if( as.integer( exists) ){ print("Exists!") }else{ dengue_model_conn(tableData, disease, conn, degree) }
RPostgres::dbDisconnect(conn)
}

