% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ws2pgdb.R
\name{stretch_delay_latent_period}
\alias{stretch_delay_latent_period}
\title{stretch_delay_latent_period() process weather information to produce stretch-delay of latent period}
\usage{
stretch_delay_latent_period(geoid, disease, type)
}
\arguments{
\item{geoid}{FIPS number from census tables (tiger files) to determine a region}

\item{disease}{String with the name of the vector-borne disease}

\item{type}{Weather variable under investigation}
}
\value{
Returns Hellow World right now
}
\description{
This function uses the provided data frame with temperature time series to determine the
completion of latent period for a given vector-borne disease with in a region
}
\note{
In progres ....
}
\examples{
ghcnd   <- 'GHCND'
geoid   <- '48489'
type    <- 'TMAX'
disease <- 'dengue'
stretch_delay_latent_period( geoid, disease, type)
}

