% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Watts}
\alias{Watts}
\title{Extrinsic Incubation period ( EIP )}
\format{A data frame with two variables: \code{time} and \code{temperature}.
\describe{
  \item{time}{time, in days}
  \item{temperature}{temperature, in Celsius}
}}
\source{
\describe{
 Watts, D.M and Burke, D.S and Harrison, B.A and Whitmire, R.E and Nisalak, A (1987)
 Effect  of the temperature on the vector efficiency of A. Aegypti for dengue 2 virus.
 The American Journal of Tropical Medicine and Hygiene, 36(1):143
}
}
\usage{
Watts
}
\description{
The duration of the EIP determined by the air temperature.
}
\keyword{datasets}

