% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_data_avg_by_date_2_pgdb.R
\name{ws_data_avg_by_date_2_pgdb}
\alias{ws_data_avg_by_date_2_pgdb}
\title{ws_data_avg_by_date_2_pgdb() retrieves the subset of ws with intersecting data}
\usage{
ws_data_avg_by_date_2_pgdb(ghcnd, geoid, type, sDate, fDate, ws_metadata)
}
\arguments{
\item{ghcnd}{String that refers to the Global Historical Climate Network (daily) dataset}

\item{geoid}{FIPS number from census tables}

\item{type}{Variable under investigation i.e. TMAX, TMIN, PRCP}

\item{sDate}{Staring date from interval to retrieve weather data}

\item{fDate}{Final   date from interval to retrieve weather data}

\item{ws_metadata}{The table name of the subset of weather stations with intersecting data}
}
\value{
Returns the name of the table that has been created. When fail, it returns 1.
}
\description{
Same as ws_data_na_2_pgdb() , retrieves the weather data from the intersecting stations. 
It is known that NOAA enlist available data for certain years, months, or days; however, 
some of this information is missing. This function replaces the missing data with the average value of the dataset. 
The example shows the use of ws_metadata_by_date_2_pgdb(), but it can be very well be replaced by ws_metadata_2_pgdb().
}
\examples{
ghcnd        <- 'GHCND'
geoid        <- '48113'
type         <- 'TMAX'
sDate       <- '2000-01-01'
fDate       <- '2001-01-01'
stations     <- as.data.frame(all_coor_ws(ghcnd, geoid, type))
ws_metadata  <- ws_metadata_by_date_2_pgdb( geoid, type, stations, sDate, fDate)
ws_data_avg_by_date_2_pgdb(ghcnd, geoid, type, sDate, fDate, ws_metadata)

}
\keyword{NOAA,}
\keyword{Postgre,}
\keyword{rnoaa}
\keyword{station,}
\keyword{weather}
