% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ws2pgdb.R
\name{ws_metadata_span_2_pgdb}
\alias{ws_metadata_span_2_pgdb}
\title{ws_metadata_span_2_pgdb Stores NOAA information into local or remote database server}
\usage{
ws_metadata_span_2_pgdb(geoid, type, stations, span)
}
\arguments{
\item{geoid}{FIPS number from census tables (tiger files)}

\item{type}{Variable under investigation i.e. TMAX, TMIN, PRCP}

\item{stations}{The universe of stations from the rnoaa API from where the new subset will be computed.}

\item{span}{This is the threshold used to limit the search .}
}
\value{
Returns the name of the new table of the subset of weather stations with intersecting data
}
\description{
This function uses the metadata returned from all_coor_ws which in turn uses the rnoaa API
to retrieve NOAA data. Metadata includes the mindate, and maxdate dates.
Given a particular variable (i.e. TMAX, TMIN, PRCP, etc), we are interested
to retrieve the number of weather stations whose time window defined by the
interval [mindate, maxdate] intersects and the length of the time window is maximum.
Searching for answers online,the solution is a hard computing problem.
It belongs to the set of problems defined as the 'Maximum Subset Intersection' problem
which are close-related to the set cover problem studied in graph theory. The following
function computes a resulting set using an heuristic described through the comments
in the body of the function. It starts with the current year and the span variable defines
the threshold for the algorithm to search for the answers.
}
\note{
Remember that all_coor_ws() returns a set of stations.
}
\examples{
ghcnd <- 'GHCND'
geoid <- '12087'
type  <- 'TMAX'
stations <- as.data.frame( all_coor_ws( ghcnd, geoid, type) )
span <- '2'
ws_metadata_span_2_pgdb( geoid, type, stations, span)
}

