% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ws2pgdb.R
\name{ws_subset_data_avg_2_pg_db}
\alias{ws_subset_data_avg_2_pg_db}
\title{ws_subset_data_avg_2_pg_db() retrieves the subset of ws with intersectin data}
\usage{
ws_subset_data_avg_2_pg_db(ghcnd, geoid, type, ws_subset_metadata)
}
\arguments{
\item{ghcnd}{String that refers to the Global Historical Climate Network (daily) dataset}

\item{geoid}{FIPS number from census tables}

\item{type}{Variable under investigation i.e. TMAX, TMIN, PRCP}

\item{ws_subset_metadata}{The table name of the subset of weather stations with intersecting data}
}
\value{
Returns the name of the table that has been created. When fail, it returns 1.
}
\description{
Same as ws_subset_data_na_2_pg_db() , retrieves the weather data from the intersecting stations.
It is known that NOAA enlist available data for certain years, months, or days; however,
some of this information is missing. This function replaces the missing data with the average value of the dataset.
The example shows the use of ws_subset_metadata_span_2_pg_db(), but it can be very well be replaced by ws_subset_metadata_2_pg_db().
}
\examples{
ghcnd <- 'GHCND'
geoid <- '48001'
type  <- 'TMAX'
stations <- as.data.frame( all_coor_ws( ghcnd, geoid, type) )
span  <- 2
ws_subset_metadata  <- ws_subset_metadata_span_2_pg_db( geoid, type, stations, span )
ws_subset_data_avg_2_pg_db( ghcnd, geoid, type, ws_subset_metadata)
}
\keyword{NOAA,}
\keyword{RPostgreSQL,}
\keyword{rgdal,}
\keyword{rnoaa}
\keyword{station,}
\keyword{weather}

