% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ws2pgdb.R
\name{ws_data_avg_2_pgdb}
\alias{ws_data_avg_2_pgdb}
\title{ws_data_avg_2_pgdb() retrieves the subset of ws with intersectin data}
\usage{
ws_data_avg_2_pgdb(ghcnd, geoid, type, ws_metadata)
}
\arguments{
\item{ghcnd}{String that refers to the Global Historical Climate Network (daily) dataset}

\item{geoid}{FIPS number from census tables}

\item{type}{Variable under investigation i.e. TMAX, TMIN, PRCP}

\item{ws_metadata}{The table name of the subset of weather stations with intersecting data}
}
\value{
Returns the name of the table that has been created. When fail, it returns 1.
}
\description{
Same as ws_data_na_2_pgdb() , retrieves the weather data from the intersecting stations.
It is known that NOAA enlist available data for certain years, months, or days; however,
some of this information is missing. This function replaces the missing data with the average value of the dataset.
The example shows the use of ws_metadata_span_2_pgdb(), but it can be very well be replaced by ws_metadata_2_pgdb().
}
\examples{
ghcnd <- 'GHCND'
geoid <- '12087'
type  <- 'TMAX'
stations <- as.data.frame( all_coor_ws( ghcnd, geoid, type) )
span <-'2'
ws_metadata  <- ws_metadata_span_2_pgdb( geoid, type, stations, span )
ws_data_avg_2_pgdb( ghcnd, geoid, type, ws_metadata)
}
\keyword{NOAA,}
\keyword{RPostgreSQL,}
\keyword{rgdal,}
\keyword{rnoaa}
\keyword{station,}
\keyword{weather}

